requirements_linux.txt
absl-py==0.9.0
actionlib==1.11.13
angles==1.9.12
astor==0.8.1
attrs==15.2.0
autobahn==0.10.3
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
bboxes @ git+https://github.com/czrcbl/bboxes@8b4fc4292362c0dfc284e3d0cc987a8a732841be
beautifulsoup4==4.4.1
bondpy==1.8.3
bzr==2.7.0
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.20
catkin-pkg==0.4.18
catkin-pkg-modules==0.4.18
catkin-tools==0.6.1
certifi==2020.4.5.1
chardet==3.0.4
cloudpickle==1.3.0
configobj==5.0.6
controller-manager==0.13.5
controller-manager-msgs==0.13.5
cryptography==1.2.3
cv-bridge==1.12.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.4.1
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.12
dynamic-reconfigure==1.5.50
ecdsa==0.13
empy==3.3.2
enum34==1.1.10
funcsigs==1.0.2
futures==3.0.5
gast==0.3.3
gazebo-plugins==2.5.19
gazebo-ros==2.5.19
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
gluoncv==0.7.0
gmpy==1.17
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.28.1
h5py==2.10.0
html5lib==0.999
httplib2==0.9.1
idna==2.9
image-geometry==1.12.8
interactive-markers==1.11.5
ipaddress==1.0.16
ipython==2.4.1
joint-state-publisher==1.12.15
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==7.3
kiwisolver==1.1.0
laser-geometry==1.6.5
launchpadlib==1.10.3
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
lxml==3.5.0
lz4==0.7.0
Magic-file-extensions==0.2
Markdown==3.1.1
matplotlib==2.2.4
mercurial==3.7.3
message-filters==1.12.14
mock==3.0.5
moveit-commander==0.9.18
moveit-ros-planning-interface==0.9.18
moveit-ros-visualization==0.9.18
mpi4py==1.3.1
mpmath==0.19
msgpack-python==0.4.6
mxnet==1.6.0
mxnet-cu100==1.5.1.post0
netifaces==0.10.4
networkx==2.2
nose==1.3.7
numpy==1.16.6
oauth==1.0.1
opencv-python==4.2.0.32
openravepy==0.9.0
osrf-pycommon==0.1.9
PAM==0.4.2
paramiko==1.16.0
pathlib==1.0.1
pexpect==4.0.1
Pillow==6.2.2
pluginlib==1.11.3
portalocker==1.7.0
protobuf==3.11.3
psutil==3.4.2
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.7
PyAssimp==3.0.1264
pycrypto==2.6.1
pydot==1.0.29
pyglet==1.1.4
Pygments==2.1
pygobject==3.20.0
pygpgme==0.3
pyliblzma==0.5.3
pymodbus==1.3.0
PyOpenGL==3.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyserial==3.0.1
Pyste==0.9.10
python-dateutil==2.4.2
python-qt-binding==0.3.7
python-snappy==0.5
pytz==2014.10
PyWavelets==1.0.3
PyYAML==3.11
qt-dotgraph==0.3.17
qt-gui==0.3.17
qt-gui-cpp==0.3.17
qt-gui-py-common==0.3.17
requests==2.23.0
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.12.14
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosdep==0.19.0
rosdep-modules==0.19.0
rosdistro==0.8.1
rosdistro-modules==0.8.1
rosgraph==1.12.14
rosinstall==0.7.8
rosinstall-generator==0.1.19
roslaunch==1.12.14
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.6
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.2.6
rospkg-modules==1.2.6
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.6
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.14
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.12.17
scikit-image==0.14.5
scipy==1.2.3
SecretStorage==2.1.3
sensor-msgs==1.12.7
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.8.1
six==1.10.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
SOAPpy==0.12.22
srdfdom==0.4.2
subprocess32==3.5.4
sympy==0.7.6.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.14
tqdm==4.46.0
trollius==2.0.1
Twisted==16.0.0
txaio==1.0.0
urdfdom-py==0.3.3
urllib3==1.25.9
vcstools==0.1.42
wadllib==1.3.2
Werkzeug==1.0.1
wrapt==1.12.1
wstool==0.1.17
wstools==0.4.3
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xacro==1.11.3
zope.interface==4.1.3
